[#43] 옵션 창 띄우는 방식 변경
기존 Instantiate, Destroy -> 변경 Instantiate, SetActive
